requirements.txt
pyserial
python-uinput
python-evdev
dataclasses-json